Flysky

+++ src/SoftIBus.cpp
@@ -29,7 +29,7 @@ void SIBus::loop(void)
     }
     last = now;
     
-    uint8_t v = stream->read();
+   uint8_t v = stream->read(); /* uint8_t v = stream->read(); */
     switch (state)
     {
       case GET_LENGTH:
@@ -62,7 +62,7 @@ void SIBus::loop(void)
 
       case GET_CHKSUMH:
         // Validate checksum
-        if (chksum == (v << 8) + lchksum)
+      if (chksum == static_cast<uint16_t>(v << 8) + lchksum)  // было if (chksum == (v << 8) + lchksum)
         {
           // Execute command - we only know command 0x40
           switch (buffer[0])
